requirements.txt
fastapi
uvicorn
psycopg2-binary
openpyxl
fpdf
python-multipart 